Add files via upload
add file from pure f(x)
Now gener all table with players(without comare)
